#661 - filters - added possibility to create theme filters
