+++ core/app.js
@@ -6,22 +6,21 @@ import config from 'config'
 
 import { sync } from 'vuex-router-sync'
 
-import { registerTheme, plugins, mixins } from 'core/lib/themes'
+import { registerTheme, plugins, mixins, filters } from 'core/lib/themes'
 import { registerExtensions } from 'core/lib/extensions'
-import * as filters from 'core/filters'
 
 import VueLazyload from 'vue-lazyload'
 import Vuelidate from 'vuelidate'
 import i18n from 'core/lib/i18n'
 
-const pluginsObject = plugins()
-Object.keys(pluginsObject).forEach(function (key) {
-  Vue.use(pluginsObject[key])
-})
-
 Vue.use(Vuelidate)
 Vue.use(VueLazyload, {
   attempt: 2
+})
+
+const pluginsObject = plugins()
+Object.keys(pluginsObject).forEach(function (key) {
+  Vue.use(pluginsObject[key])
 })
 
 const mixinsObject = mixins()
@@ -29,9 +28,9 @@ Object.keys(mixinsObject).forEach(function (key) {
   Vue.mixin(mixinsObject[key])
 })
 
-// Register global filters
-Object.keys(filters).forEach(key => {
-  Vue.filter(key, filters[key])
+const filtersObject = filters()
+Object.keys(filtersObject).forEach(key => {
+  Vue.filter(key, filtersObject[key])
 })
 
 export function createApp () {
